test(productsModule): products module tests have been adjusted
the products module test have been updated, and covered increased

 this commit is connected to #5
